style: (#178) usecase 패키지로 이동
